requirements.txt
tensorflow==2.15.0
streamlit==1.31.0
Pillow==10.2.0
numpy==1.26.4